feat: Simplify the runner class
