+++ bluebird-api/bluebird_api/routes.py
@@ -5,27 +5,17 @@ BluebirdATC, including loading which is implementation dependent.
 
 import asyncio
 
-from fastapi import APIRouter, Depends, Request
+from bluebird_dt.simulator import Simulator
+from fastapi import APIRouter
 
-from bluebird_api.models import RunnerStore
 from bluebird_api.routers.core import background_tasks
-from bluebird_api.runner import Runner
+from bluebird_api.runner import Runner, RunnerStore
 
 from .routers import (
     core_router,
 )
 
-
-async def simulator(request: Request):
-    """
-    The simulator dependency, available to endpoints using the RunnerDep dependency, provides access to the runner.
-    """
-
-    request.state.runner = RunnerStore.current_runner
-
-
-router = APIRouter(dependencies=[Depends(simulator)])
-
+router = APIRouter()
 router.include_router(core_router)
 
 
@@ -34,11 +24,10 @@ async def load(category: str, scenario_name: str) -> bool:  # noqa: ARG001
     """
     End any existing run, then create a new Runner and load a given simulator scenario.
     """
-
     if RunnerStore.current_runner is not None:
         await RunnerStore.current_runner.delete()
 
-    RunnerStore.current_runner = Runner(category, scenario_name)
+    RunnerStore.current_runner = Runner(Simulator.from_category(category, scenario_name))
 
     # start the task
     task = asyncio.create_task(RunnerStore.current_runner.run_main())
